// src/app/_ngrx-store/enums/article-action-types.enum.ts
export enum ArticleActionTypes {
    Get = '[ARTICLE] Get',
    Get_Success = '[ARTICLE] Get success',
    Get_Error = '[ARTICLE] Get Error',

    Add = '[ARTICLE] Add',
    Add_Success = '[ARTICLE] Add Success',
    Add_Error = '[ARTICLE] Add Error',

    Delete = '[ARTICLE] Delete',
    Delete_Success = '[ARTICLE] Delete Succes',
    Delete_Error = '[ARTICLE] Delete Error',
    
    Edit = '[ARTICLE] Edit',
    Edit_Success = '[ARTICLE] Edit Suecces',
    Edit_Error = '[ARTICLE] Edit Error',
  }